Simplify bbl detection
Instead of reading only a tex file and loading dependencies dynamically, let’s be naive for now and
parse all '*.tex' and '*.bbl' content. This avoids the need to find latex aliases and to find the
latex document structure (i.e. we don't need to run the parsing on the 'root' document).